Enum script_traits::CompositorEvent

pub enum CompositorEvent {
    ResizeEvent(WindowSizeData, WindowSizeType),
    MouseButtonEvent(MouseEventType, MouseButton, Point2D<f32>),
    MouseMoveEvent(Option<Point2D<f32>>),
    TouchEvent(TouchEventType, TouchId, Point2D<f32>),
    TouchpadPressureEvent(Point2D<f32>, f32, TouchpadPressurePhase),
    KeyEvent(Key, KeyState, KeyModifiers),
}

Events from the compositor that the script thread needs to know about

Variants

ResizeEvent

The window was resized.

MouseButtonEvent

A mouse button state changed.

MouseMoveEvent

The mouse was moved over a point (or was moved out of the recognizable region).

TouchEvent

A touch event was generated with a touch ID and location.

TouchpadPressureEvent

Touchpad pressure event

KeyEvent

A key was pressed.