Enum script_traits::ConstellationControlMsg

pub enum ConstellationControlMsg {
    AttachLayout(NewLayoutInfo),
    Resize(PipelineId, WindowSizeData, WindowSizeType),
    ResizeInactive(PipelineId, WindowSizeData),
    ExitPipeline(PipelineId),
    ExitScriptThread,
    SendEvent(PipelineId, CompositorEvent),
    Viewport(PipelineId, Rect<f32>),
    SetScrollState(PipelineId, Vec<(UntrustedNodeAddress, Point2D<f32>)>),
    GetTitle(PipelineId),
    Freeze(PipelineId),
    Thaw(PipelineId),
    ChangeFrameVisibilityStatus(PipelineId, bool),
    NotifyVisibilityChange(PipelineId, FrameId, bool),
    Navigate(PipelineId, FrameId, LoadData, bool),
    MozBrowserEvent(PipelineId, Option<FrameId>, MozBrowserEvent),
    UpdatePipelineId(PipelineId, FrameId, PipelineId),
    FocusIFrame(PipelineId, FrameId),
    WebDriverScriptCommand(PipelineId, WebDriverScriptCommand),
    TickAllAnimations(PipelineId),
    TransitionEnd(UnsafeNode, String, f64),
    WebFontLoaded(PipelineId),
    DispatchFrameLoadEvent {
        target: FrameId,
        parent: PipelineId,
        child: PipelineId,
    },
    FramedContentChanged(PipelineId, FrameId),
    ReportCSSError(PipelineId, String, usize, usize, String),
    Reload(PipelineId),
}

Messages sent from the constellation or layout to the script thread.

Variants
AttachLayout(NewLayoutInfo)

Gives a channel and ID to a layout thread, as well as the ID of that layout's parent

Resize(PipelineId, WindowSizeData, WindowSizeType)

Window resized. Sends a DOM event eventually, but first we combine events.

ResizeInactive(PipelineId, WindowSizeData)

Notifies script that window has been resized but to not take immediate action.

ExitPipeline(PipelineId)

Notifies the script that a pipeline should be closed.

ExitScriptThread

Notifies the script that the whole thread should be closed.

SendEvent(PipelineId, CompositorEvent)

Sends a DOM event.

Viewport(PipelineId, Rect<f32>)

Notifies script of the viewport.

SetScrollState(PipelineId, Vec<(UntrustedNodeAddress, Point2D<f32>)>)

Notifies script of a new set of scroll offsets.

GetTitle(PipelineId)

Requests that the script thread immediately send the constellation the title of a pipeline.

Freeze(PipelineId)

Notifies script thread to suspend all its timers

Thaw(PipelineId)

Notifies script thread to resume all its timers

ChangeFrameVisibilityStatus(PipelineId, bool)

Notifies script thread whether frame is visible

NotifyVisibilityChange(PipelineId, FrameId, bool)

Notifies script thread that frame visibility change is complete PipelineId is for the parent, FrameId is for the actual frame.

Navigate(PipelineId, FrameId, LoadData, bool)

Notifies script thread that a url should be loaded in this iframe. PipelineId is for the parent, FrameId is for the actual frame.

MozBrowserEvent(PipelineId, Option<FrameId>, MozBrowserEvent)

Requests the script thread forward a mozbrowser event to an iframe it owns, or to the window if no child frame id is provided.

UpdatePipelineId(PipelineId, FrameId, PipelineId)

Updates the current pipeline ID of a given iframe. First PipelineId is for the parent, second is the new PipelineId for the frame.

FocusIFrame(PipelineId, FrameId)

Set an iframe to be focused. Used when an element in an iframe gains focus. PipelineId is for the parent, FrameId is for the actual frame.

WebDriverScriptCommand(PipelineId, WebDriverScriptCommand)

Passes a webdriver command to the script thread for execution

TickAllAnimations(PipelineId)

Notifies script thread that all animations are done

TransitionEnd(UnsafeNode, String, f64)

Notifies the script thread of a transition end

WebFontLoaded(PipelineId)

Notifies the script thread that a new Web font has been loaded, and thus the page should be reflowed.

DispatchFrameLoadEvent

Cause a load event to be dispatched at the appropriate frame element.

Fields

target: FrameId

The frame that has been marked as loaded.

parent: PipelineId

The pipeline that contains a frame loading the target pipeline.

child: PipelineId

The pipeline that has completed loading.

FramedContentChanged(PipelineId, FrameId)

Notifies a parent pipeline that one of its child frames is now active. PipelineId is for the parent, FrameId is the child frame.

ReportCSSError(PipelineId, String, usize, usize, String)

Report an error from a CSS parser for the given pipeline

Reload(PipelineId)

Reload the given page.