Enum script_traits::LayoutControlMsg

pub enum LayoutControlMsg {
    ExitNow,
    GetCurrentEpoch(IpcSender<Epoch>),
    TickAnimations,
    SetVisibleRects(Vec<(LayerId, Rect<Au>)>),
    SetStackingContextScrollStates(Vec<StackingContextScrollState>),
    GetWebFontLoadState(IpcSender<bool>),
}

Messages sent to the layout thread from the constellation and/or compositor.

Variants

ExitNow

Requests that this layout thread exit.

GetCurrentEpoch

Requests the current epoch (layout counter) from this layout.

TickAnimations

Asks layout to run another step in its animation.

SetVisibleRects

Informs layout as to which regions of the page are visible.

SetStackingContextScrollStates

Tells layout about the new scrolling offsets of each scrollable stacking context.

GetWebFontLoadState

Requests the current load state of Web fonts. true is returned if fonts are still loading and false is returned if all fonts have loaded.