Enum script_traits::MozBrowserEvent

pub enum MozBrowserEvent {
    AsyncScroll,
    Close,
    ContextMenu,
    Error(MozBrowserErrorType, String, String),
    IconChange(String, String, String),
    Connected,
    LoadEnd,
    LoadStart,
    LocationChange(String, bool, bool),
    OpenTab(String),
    OpenWindow(String, Option<String>, Option<String>),
    SecurityChange(HttpsState),
    ShowModalPrompt(String, String, String, String),
    TitleChange(String),
    UsernameAndPasswordRequired,
    OpenSearch,
    VisibilityChange(bool),
}

The events fired in a Browser API context (<iframe mozbrowser>)

Variants
AsyncScroll

Sent when the scroll position within a browser <iframe> changes.

Close

Sent when window.close() is called within a browser <iframe>.

ContextMenu

Sent when a browser <iframe> tries to open a context menu. This allows handling <menuitem> element available within the browser <iframe>'s content.

Error(MozBrowserErrorType, String, String)

Sent when an error occurred while trying to load content within a browser <iframe>. Includes a human-readable description, and a machine-readable report.

IconChange(String, String, String)

Sent when the favicon of a browser <iframe> changes.

Connected

Sent when the browser <iframe> has reached the server.

LoadEnd

Sent when the browser <iframe> has finished loading all its assets.

LoadStart

Sent when the browser <iframe> starts to load a new page.

LocationChange(String, bool, bool)

Sent when a browser <iframe>'s location changes.

OpenTab(String)

Sent when a new tab is opened within a browser <iframe> as a result of the user issuing a command to open a link target in a new tab (for example ctrl/cmd + click.) Includes the URL.

OpenWindow(String, Option<String>, Option<String>)

Sent when a new window is opened within a browser <iframe>. Includes the URL, target browsing context name, and features.

SecurityChange(HttpsState)

Sent when the SSL state changes within a browser <iframe>.

ShowModalPrompt(String, String, String, String)

Sent when alert(), confirm(), or prompt() is called within a browser <iframe>.

TitleChange(String)

Sent when the document.title changes within a browser <iframe>.

UsernameAndPasswordRequired

Sent when an HTTP authentification is requested.

OpenSearch

Sent when a link to a search engine is found.

VisibilityChange(bool)

Sent when visibility state changes.

Methods

impl MozBrowserEvent

fn name(&self) -> &'static str

Get the name of the event as a & str