script_traits

Enum ScriptThreadMessage

pub enum ScriptThreadMessage {

    SpawnPipeline(NewPipelineInfo),
    StopDelayingLoadEventsMode(PipelineId),
    Resize(PipelineId, ViewportDetails, WindowSizeType),
    ThemeChange(PipelineId, Theme),
    ResizeInactive(PipelineId, ViewportDetails),
    ExitFullScreen(PipelineId),
    UnloadDocument(PipelineId),
    ExitPipeline(WebViewId, PipelineId, DiscardBrowsingContext),
    ExitScriptThread,
    SendInputEvent(WebViewId, PipelineId, ConstellationInputEvent),
    RefreshCursor(PipelineId),
    GetTitle(PipelineId),
    GetDocumentOrigin(PipelineId, GenericSender<Option<OriginSnapshot>>),
    GetInternalAncestorOriginObjectsList(PipelineId, GenericSender<Option<Vec<ImmutableOrigin>>>),
    SetDocumentActivity(PipelineId, DocumentActivity),
    SetThrottled(WebViewId, PipelineId, bool),
    SetThrottledInContainingIframe(WebViewId, PipelineId, BrowsingContextId, bool),
    NavigateIframe(PipelineId, BrowsingContextId, LoadData, NavigationHistoryBehavior, TargetSnapshotParams),
    PostMessage {
        target: PipelineId,
        source_webview: WebViewId,
        source_with_ancestry: Vec<BrowsingContextId>,
        target_origin: Option<ImmutableOrigin>,
        source_origin: ImmutableOrigin,
        data: Box<StructuredSerializedData>,
    },
    UpdatePipelineId(PipelineId, BrowsingContextId, WebViewId, PipelineId, UpdatePipelineIdReason),
    UpdateHistoryState(PipelineId, Option<HistoryStateId>, ServoUrl),
    RemoveHistoryStates(PipelineId, Vec<HistoryStateId>),
    FocusDocumentAsPartOfFocusingSteps(PipelineId, FocusSequenceNumber, Option<BrowsingContextId>),
    UnfocusDocumentAsPartOfFocusingSteps(PipelineId, FocusSequenceNumber),
    FocusDocument(PipelineId, RemoteFocusOperation),
    WebDriverScriptCommand(PipelineId, WebDriverScriptCommand),
    TickAllAnimations(Vec<WebViewId>),
    WebFontLoadFinished(PipelineId, WebFontLoadEvent),
    DispatchIFrameLoadEvent {
        target: BrowsingContextId,
        parent: PipelineId,
        child: PipelineId,
    },
    DispatchStorageEvent(PipelineId, WebStorageType, ServoUrl, Option<String>, Option<String>, Option<String>),
    ReportCSSError(PipelineId, String, u32, u32, String),
    Reload(PipelineId),
    PaintMetric(PipelineId, ProgressiveWebMetricType, CrossProcessInstant, bool),
    MediaSessionAction(PipelineId, MediaSessionActionType),
    SetWebGPUPort(GenericReceiver<WebGPUMsg>),
    SetScrollStates(PipelineId, ScrollStateUpdate),
    EvaluateJavaScript(WebViewId, PipelineId, JavaScriptEvaluationId, String),
    SendImageKeysBatch(PipelineId, Vec<ImageKey>),
    PreferencesUpdated(Vec<(String, PrefValue)>),
    NoLongerWaitingOnAsychronousImageUpdates(PipelineId),
    ForwardKeyboardScroll(PipelineId, KeyboardScroll),
    RequestScreenshotReadiness(WebViewId, PipelineId),
    EmbedderControlResponse(EmbedderControlId, EmbedderControlResponse),
    SetUserContents(UserContentManagerId, UserContents),
    DestroyUserContentManager(UserContentManagerId),
    UpdatePinchZoomInfos(PipelineId, PinchZoomInfos),
    SetAccessibilityActive(PipelineId, bool, Epoch),
    TriggerGarbageCollection,

}

Messages sent to the ScriptThread event loop from the Constellation, Paint, and (for now) Layout.

Variants

SpawnPipeline(NewPipelineInfo)

Span a new Pipeline in this ScriptThread and start fetching the contents according to the provided LoadData. This will ultimately create a Window and all associated data structures such as Layout in the ScriptThread.

StopDelayingLoadEventsMode(PipelineId)

Takes the associated window proxy out of “delaying-load-events-mode”, used if a scheduled navigated was refused by the embedder. https://html.spec.whatwg.org/multipage/#delaying-load-events-mode

Resize(PipelineId, ViewportDetails, WindowSizeType)

Window resized. Sends a DOM event eventually, but first we combine events.

ThemeChange(PipelineId, Theme)

Theme changed.

ResizeInactive(PipelineId, ViewportDetails)

Notifies script that window has been resized but to not take immediate action.

ExitFullScreen(PipelineId)

Window switched from fullscreen mode.

UnloadDocument(PipelineId)

Notifies the script that the document associated with this pipeline should ‘unload’.

ExitPipeline(WebViewId, PipelineId, DiscardBrowsingContext)

Notifies the script that a pipeline should be closed.

ExitScriptThread

Notifies the script that the whole thread should be closed.

SendInputEvent(WebViewId, PipelineId, ConstellationInputEvent)

Sends a DOM event.

RefreshCursor(PipelineId)

Request that the given pipeline refresh the cursor by doing a hit test at the most recently hovered cursor position and resetting the cursor. This happens after a display list update is rendered.

GetTitle(PipelineId)

Requests that the script thread immediately send the constellation the title of a pipeline.

GetDocumentOrigin(PipelineId, GenericSender<Option<OriginSnapshot>>)

Retrieve the origin of a document for a pipeline, in case a child needs to retrieve the origin of a parent in a different script thread.

GetInternalAncestorOriginObjectsList(PipelineId, GenericSender<Option<Vec<ImmutableOrigin>>>)

Retrieve the internal ancestor origin objects list of a document for a pipeline, in case a child needs to retrieve the origin of a parent in a different script thread.

SetDocumentActivity(PipelineId, DocumentActivity)

Notifies script thread of a change to one of its document’s activity

SetThrottled(WebViewId, PipelineId, bool)

Set whether to use less resources by running timers at a heavily limited rate.

SetThrottledInContainingIframe(WebViewId, PipelineId, BrowsingContextId, bool)

Notify the containing iframe (in PipelineId) that the nested browsing context (BrowsingContextId) is throttled.

NavigateIframe(PipelineId, BrowsingContextId, LoadData, NavigationHistoryBehavior, TargetSnapshotParams)

Notifies script thread that a url should be loaded in this iframe. PipelineId is for the parent, BrowsingContextId is for the nested browsing context

PostMessage

Post a message to a given window.

Fields

target: PipelineId

The target of the message.

source_webview: WebViewId

The webview associated with the source pipeline.

source_with_ancestry: Vec<BrowsingContextId>

The ancestry of browsing context associated with the source, starting with the source itself.

target_origin: Option<ImmutableOrigin>

The expected origin of the target.

source_origin: ImmutableOrigin

The source origin of the message. https://html.spec.whatwg.org/multipage/#dom-messageevent-origin

data: Box<StructuredSerializedData>

The data to be posted.

UpdatePipelineId(PipelineId, BrowsingContextId, WebViewId, PipelineId, UpdatePipelineIdReason)

Updates the current pipeline ID of a given iframe. First PipelineId is for the parent, second is the new PipelineId for the frame.

UpdateHistoryState(PipelineId, Option<HistoryStateId>, ServoUrl)

Updates the history state and url of a given pipeline.

RemoveHistoryStates(PipelineId, Vec<HistoryStateId>)

Removes inaccesible history states.

FocusDocumentAsPartOfFocusingSteps(PipelineId, FocusSequenceNumber, Option<BrowsingContextId>)

Focus a Document as part of the focusing steps which focuses all parent Documents of a newly focused <iframe>. Note that this is not used for the Document and Element that is gaining focus as that is handled locally in the originating ScriptThread.

UnfocusDocumentAsPartOfFocusingSteps(PipelineId, FocusSequenceNumber)

Unfocus a Document as part of the focusing steps which unfocuses all parent Documents of an <iframe> losing focus. This does not do anything for a top-level Document, which can never lose focus (apart from losing system focus, which is a separate concept).

FocusDocument(PipelineId, RemoteFocusOperation)

Focus a Document and run the focusing steps. This is used in two situations:

When calling the DOM focus() API on a remote Window as well as from WebDriver. The difference between this and FocusDocumentAsPartOfFocusingSteps is that this version actually does run the focusing steps and may result in blur and focus events firing up the frame tree.

When doing sequential focus navigation into and out of frames.

WebDriverScriptCommand(PipelineId, WebDriverScriptCommand)

Passes a webdriver command to the script thread for execution

TickAllAnimations(Vec<WebViewId>)

Notifies script thread that all animations are done

WebFontLoadFinished(PipelineId, WebFontLoadEvent)

Notifies the script thread that a web font has finished loading.

This is sent if either the web font loaded successfully, or to notify the script thread that it should try to resolve document.fonts.ready because the font was the last one loading.

DispatchIFrameLoadEvent

Cause a load event to be dispatched at the appropriate iframe element.

Fields

target: BrowsingContextId

The frame that has been marked as loaded.

parent: PipelineId

The pipeline that contains a frame loading the target pipeline.

child: PipelineId

The pipeline that has completed loading.

DispatchStorageEvent(PipelineId, WebStorageType, ServoUrl, Option<String>, Option<String>, Option<String>)

Cause a storage event to be dispatched at the appropriate window. The strings are key, old value and new value.

ReportCSSError(PipelineId, String, u32, u32, String)

Report an error from a CSS parser for the given pipeline

Reload(PipelineId)

Reload the given page.

PaintMetric(PipelineId, ProgressiveWebMetricType, CrossProcessInstant, bool)

Notifies the script thread about a new recorded paint metric.

MediaSessionAction(PipelineId, MediaSessionActionType)

Notifies the media session about a user requested media session action.

SetWebGPUPort(GenericReceiver<WebGPUMsg>)

Notifies script thread that WebGPU server has started

SetScrollStates(PipelineId, ScrollStateUpdate)

Paint scrolled and is updating the scroll states of the nodes in the given pipeline via the Constellation.

EvaluateJavaScript(WebViewId, PipelineId, JavaScriptEvaluationId, String)

Evaluate the given JavaScript and return a result via a corresponding message to the Constellation.

SendImageKeysBatch(PipelineId, Vec<ImageKey>)

A new batch of keys for the image cache for the specific pipeline.

PreferencesUpdated(Vec<(String, PrefValue)>)

Preferences were updated in the parent process.

NoLongerWaitingOnAsychronousImageUpdates(PipelineId)

Notify the ScriptThread that the Servo renderer is no longer waiting on asynchronous image uploads for the given Pipeline. These are mainly used by canvas to perform uploads while the display list is being built.

ForwardKeyboardScroll(PipelineId, KeyboardScroll)

Forward a keyboard scroll operation from an <iframe> to a parent pipeline.

RequestScreenshotReadiness(WebViewId, PipelineId)

Request readiness for a screenshot from the given pipeline. The pipeline will respond when it is ready to take the screenshot or will not be able to take it in the future.

EmbedderControlResponse(EmbedderControlId, EmbedderControlResponse)

A response to a request to show an embedder user interface control.

SetUserContents(UserContentManagerId, UserContents)

Set the UserContents for the given UserContentManagerId. A ScriptThread can host many WebViews which share the same UserContentManager. Only documents loaded after the processing of this message will observe the new UserContents of the specified UserContentManagerId.

DestroyUserContentManager(UserContentManagerId)

Release all data for the given UserContentManagerId from the ScriptThread’s user_contents_for_manager_id map.

UpdatePinchZoomInfos(PipelineId, PinchZoomInfos)

Update the pinch zoom details of a pipeline. Each Window stores a VisualViewport DOM instance that gets updated according to the changes from the `Compositor``.

SetAccessibilityActive(PipelineId, bool, Epoch)

Activate or deactivate accessibility features for the given pipeline, assuming it represents a document.

Why only one pipeline? In the Servo API, accessibility is activated on a per-webview basis, and webviews have a simple one-to-many mapping to pipelines that represent documents. But those pipelines run in script threads, which complicates things: the pipelines in a webview may be split across multiple script threads, and the pipelines in a script thread may belong to multiple webviews. So the simplest approach is to activate it for one pipeline at a time.

TriggerGarbageCollection

Force a garbage collection in this script thread.