Enum script_traits::ScriptToCompositorMsg

pub enum ScriptToCompositorMsg {
    ScrollFragmentPoint(PipelineId, LayerId, Point2D<f32>, bool),
    SetTitle(PipelineId, Option<String>),
    SendKeyEvent(Key, KeyState, KeyModifiers),
    GetClientWindow(IpcSender<(Size2D<u32>, Point2D<i32>)>),
    MoveTo(Point2D<i32>),
    ResizeTo(Size2D<u32>),
    TouchEventProcessed(EventResult),
    GetScrollOffset(PipelineId, LayerId, IpcSender<Point2D<f32>>),
    Exit,
    Exited,
}

Messages sent from the script thread to the compositor

Variants

ScrollFragmentPoint

Scroll a page in a window

SetTitle

Set title of current page https://html.spec.whatwg.org/multipage/#document.title

SendKeyEvent

Send a key event

GetClientWindow

Get Window Informations size and position

MoveTo

Move the window to a point

ResizeTo

Resize the window to size

TouchEventProcessed

Script has handled a touch event, and either prevented or allowed default actions.

GetScrollOffset

Get Scroll Offset

Exit

Requests that the compositor shut down.

Exited

Allow the compositor to free script-specific resources.