Enum script_traits::WebDriverCommandMsg

pub enum WebDriverCommandMsg {
    GetWindowSize(PipelineId, IpcSender<WindowSizeData>),
    LoadUrl(PipelineId, LoadData, IpcSender<LoadStatus>),
    Refresh(PipelineId, IpcSender<LoadStatus>),
    ScriptCommand(PipelineId, WebDriverScriptCommand),
    SendKeys(PipelineId, Vec<(Key, KeyModifiers, KeyState)>),
    SetWindowSize(PipelineId, Size2D<u32>, IpcSender<WindowSizeData>),
    TakeScreenshot(PipelineId, IpcSender<Option<Image>>),
}

Messages to the constellation originating from the WebDriver server.

Variants
GetWindowSize(PipelineId, IpcSender<WindowSizeData>)

Get the window size.

LoadUrl(PipelineId, LoadData, IpcSender<LoadStatus>)

Load a URL in the pipeline with the given ID.

Refresh(PipelineId, IpcSender<LoadStatus>)

Refresh the pipeline with the given ID.

ScriptCommand(PipelineId, WebDriverScriptCommand)

Pass a webdriver command to the script thread of the pipeline with the given ID for execution.

SendKeys(PipelineId, Vec<(Key, KeyModifiers, KeyState)>)

Act as if keys were pressed in the pipeline with the given ID.

SetWindowSize(PipelineId, Size2D<u32>, IpcSender<WindowSizeData>)

Set the window size.

TakeScreenshot(PipelineId, IpcSender<Option<Image>>)

Take a screenshot of the window, if the pipeline with the given ID is the root pipeline.