Enum script_traits::script_msg::LogEntry

pub enum LogEntry {
    Panic(String, String),
    Error(String),
    Warn(String),
}

A log entry reported to the constellation We don't report all log entries, just serious ones. We need a separate type for this because LogLevel isn't serializable.

Variants
Panic(String, String)

Panic, with a reason and backtrace

Error(String)

Error, with a reason

Warn(String)

warning, with a reason