Enum script_traits::script_msg::ScriptMsg

pub enum ScriptMsg {
    BroadcastStorageEvent(PipelineId, StorageType, ServoUrl, Option<String>, Option<String>, Option<String>),
    ChangeRunningAnimationsState(PipelineId, AnimationState),
    CreateCanvasPaintThread(Size2D<i32>, IpcSender<IpcSender<CanvasMsg>>),
    CreateWebGLPaintThread(Size2D<i32>, GLContextAttributes, IpcSender<Result<(IpcSender<CanvasMsg>, GLLimits), String>>),
    Focus(PipelineId),
    ForwardEvent(PipelineId, CompositorEvent),
    GetClipboardContents(IpcSender<String>),
    HeadParsed,
    LoadComplete(PipelineId),
    LoadUrl(PipelineId, LoadData, bool),
    MozBrowserEvent(PipelineId, PipelineId, MozBrowserEvent),
    TraverseHistory(Option<PipelineId>, TraversalDirection),
    JointSessionHistoryLength(PipelineId, IpcSender<u32>),
    NewFavicon(ServoUrl),
    NodeStatus(Option<String>),
    RemoveIFrame(PipelineId, Option<IpcSender<()>>),
    SetVisible(PipelineId, bool),
    VisibilityChangeComplete(PipelineId, bool),
    ScriptLoadedURLInIFrame(IFrameLoadInfo),
    SetClipboardContents(String),
    ActivateDocument(PipelineId),
    SetDocumentState(PipelineId, DocumentState),
    SetFinalUrl(PipelineId, ServoUrl),
    Alert(PipelineId, String, IpcSender<bool>),
    ScrollFragmentPoint(PipelineId, Point2D<f32>, bool),
    SetTitle(PipelineId, Option<String>),
    SendKeyEvent(Option<char>, Key, KeyState, KeyModifiers),
    GetClientWindow(IpcSender<(Size2D<u32>, Point2D<i32>)>),
    MoveTo(Point2D<i32>),
    ResizeTo(Size2D<u32>),
    TouchEventProcessed(EventResult),
    LogEntry(Option<PipelineId>, Option<String>, LogEntry),
    PipelineExited(PipelineId),
    ForwardDOMMessage(DOMMessage, ServoUrl),
    RegisterServiceWorker(ScopeThings, ServoUrl),
    Exit,
}

Messages from the script to the constellation.

Variants
BroadcastStorageEvent(PipelineId, StorageType, ServoUrl, Option<String>, Option<String>, Option<String>)

Broadcast a storage event to every same-origin pipeline. The strings are key, old value and new value.

ChangeRunningAnimationsState(PipelineId, AnimationState)

Indicates whether this pipeline is currently running animations.

CreateCanvasPaintThread(Size2D<i32>, IpcSender<IpcSender<CanvasMsg>>)

Requests that a new 2D canvas thread be created. (This is done in the constellation because 2D canvases may use the GPU and we don't want to give untrusted content access to the GPU.)

CreateWebGLPaintThread(Size2D<i32>, GLContextAttributes, IpcSender<Result<(IpcSender<CanvasMsg>, GLLimits), String>>)

Requests that a new WebGL thread be created. (This is done in the constellation because WebGL uses the GPU and we don't want to give untrusted content access to the GPU.)

Focus(PipelineId)

Notifies the constellation that this frame has received focus.

ForwardEvent(PipelineId, CompositorEvent)

Forward an event that was sent to the parent window.

GetClipboardContents(IpcSender<String>)

Requests that the constellation retrieve the current contents of the clipboard

HeadParsed

LoadComplete(PipelineId)

All pending loads are complete, and the load event for this pipeline has been dispatched.

LoadUrl(PipelineId, LoadData, bool)

A new load has been requested, with an option to replace the current entry once loaded instead of adding a new entry.

MozBrowserEvent(PipelineId, PipelineId, MozBrowserEvent)

Dispatch a mozbrowser event to the parent of this pipeline. The first PipelineId is for the parent, the second is for the originating pipeline.

TraverseHistory(Option<PipelineId>, TraversalDirection)

HTMLIFrameElement Forward or Back traversal.

JointSessionHistoryLength(PipelineId, IpcSender<u32>)

Gets the length of the joint session history from the constellation.

NewFavicon(ServoUrl)

Favicon detected

NodeStatus(Option<String>)

Status message to be displayed in the chrome, eg. a link URL on mouseover.

RemoveIFrame(PipelineId, Option<IpcSender<()>>)

Notification that this iframe should be removed.

SetVisible(PipelineId, bool)

Change pipeline visibility

VisibilityChangeComplete(PipelineId, bool)

Notifies constellation that an iframe's visibility has been changed.

ScriptLoadedURLInIFrame(IFrameLoadInfo)

A load has been requested in an IFrame.

SetClipboardContents(String)

Requests that the constellation set the contents of the clipboard

ActivateDocument(PipelineId)

Mark a new document as active

SetDocumentState(PipelineId, DocumentState)

Set the document state for a pipeline (used by screenshot / reftests)

SetFinalUrl(PipelineId, ServoUrl)

Update the pipeline Url, which can change after redirections.

Alert(PipelineId, String, IpcSender<bool>)

Check if an alert dialog box should be presented

ScrollFragmentPoint(PipelineId, Point2D<f32>, bool)

Scroll a page in a window

SetTitle(PipelineId, Option<String>)

Set title of current page https://html.spec.whatwg.org/multipage/#document.title

SendKeyEvent(Option<char>, Key, KeyState, KeyModifiers)

Send a key event

GetClientWindow(IpcSender<(Size2D<u32>, Point2D<i32>)>)

Get Window Informations size and position

MoveTo(Point2D<i32>)

Move the window to a point

ResizeTo(Size2D<u32>)

Resize the window to size

TouchEventProcessed(EventResult)

Script has handled a touch event, and either prevented or allowed default actions.

LogEntry(Option<PipelineId>, Option<String>, LogEntry)

A log entry, with the pipeline id and thread name

PipelineExited(PipelineId)

Notifies the constellation that this pipeline has exited.

ForwardDOMMessage(DOMMessage, ServoUrl)

Send messages from postMessage calls from serviceworker to constellation for storing in service worker manager

RegisterServiceWorker(ScopeThings, ServoUrl)

Store the data required to activate a service worker for the given scope

Exit

Requests that the compositor shut down.