Struct script_traits::NewLayoutInfo

pub struct NewLayoutInfo {
    pub parent_info: Option<(PipelineId, FrameType)>,
    pub new_pipeline_id: PipelineId,
    pub frame_id: FrameId,
    pub load_data: LoadData,
    pub window_size: Option<WindowSizeData>,
    pub pipeline_port: IpcReceiver<LayoutControlMsg>,
    pub content_process_shutdown_chan: Option<IpcSender<()>>,
    pub layout_threads: usize,
}

The initial data required to create a new layout attached to an existing script thread.

Fields
parent_info: Option<(PipelineId, FrameType)>

The ID of the parent pipeline and frame type, if any. If None, this is a root pipeline.

new_pipeline_id: PipelineId

Id of the newly-created pipeline.

frame_id: FrameId

Id of the frame associated with this pipeline.

load_data: LoadData

Network request data which will be initiated by the script thread.

window_size: Option<WindowSizeData>

Information about the initial window size.

pipeline_port: IpcReceiver<LayoutControlMsg>

A port on which layout can receive messages from the pipeline.

content_process_shutdown_chan: Option<IpcSender<()>>

A shutdown channel so that layout can tell the content process to shut down when it's done.

layout_threads: usize

Number of threads to use for layout.