Crate script_traits

This module contains traits in script used generically in the rest of Servo. The traits are here instead of in script so that these modules won’t have to depend on script.

Structs

ConstellationInputEvent

Input events from the embedder that are sent via the Constellation`` to the ScriptThread`.

DrawAPaintImageResult

The result of executing paint code: the image together with any image URLs that need to be loaded.

InitialScriptState

All of the information necessary to create a new [ScriptThread] for a new [EventLoop].

NewPipelineInfo

The initial data required to create a new Pipeline attached to an existing ScriptThread.

Enums

DiscardBrowsingContext

When a pipeline is closed, should its browsing context be discarded too?

DocumentActivity

Is a document fully active, active or inactive? A document is active if it is the current active document in its session history, it is fuly active if it is active and all of its ancestors are active, and it is inactive otherwise.

DocumentState

Used to determine if a script has any pending asynchronous activity.

PaintWorkletError

Errors from executing a paint worklet

ProgressiveWebMetricType

Type of recorded progressive web metric

ScriptThreadMessage

Messages sent to the ScriptThread event loop from the Constellation, Paint, and (for now) Layout.

UpdatePipelineIdReason

The reason why the pipeline id of an iframe is being updated.

Traits

Painter

Execute paint code in the worklet thread pool.