Crate script_traits

Structs

NewLayoutInfo

OpaqueScriptLayoutChannel
An opaque wrapper around script<->layout channels to avoid leaking message types into crates that don't need to know about them.

ScriptControlChan
Encapsulates external communication with the script task.

UntrustedNodeAddress
The address of a node. Layout sends these back. They must be validated via from_untrusted_node_address before they can be used, because we do not trust layout.

Enums

CompositorEvent
Events from the compositor that the script task needs to know about

ConstellationControlMsg
Messages sent from the constellation to the script task

Traits

ScriptTaskFactory