Enum script_traits::ConstellationControlMsg

pub enum ConstellationControlMsg {
    AttachLayout(NewLayoutInfo),
    Resize(PipelineId, WindowSizeData, WindowSizeType),
    ResizeInactive(PipelineId, WindowSizeData),
    ExitPipeline(PipelineId),
    SendEvent(PipelineId, CompositorEvent),
    Viewport(PipelineId, Rect<f32>),
    SetScrollState(PipelineId, Vec<(UntrustedNodeAddress, Point2D<f32>)>),
    GetTitle(PipelineId),
    Freeze(PipelineId),
    Thaw(PipelineId),
    ChangeFrameVisibilityStatus(PipelineId, bool),
    NotifyVisibilityChange(PipelineId, PipelineId, bool),
    Navigate(PipelineId, SubpageId, LoadData),
    MozBrowserEvent(PipelineId, SubpageId, MozBrowserEvent),
    UpdateSubpageId(PipelineId, SubpageId, SubpageId, PipelineId),
    FocusIFrame(PipelineId, SubpageId),
    WebDriverScriptCommand(PipelineId, WebDriverScriptCommand),
    TickAllAnimations(PipelineId),
    WebFontLoaded(PipelineId),
    DispatchFrameLoadEvent {
        target: PipelineId,
        parent: PipelineId,
    },
    FramedContentChanged(PipelineId, SubpageId),
    ReportCSSError(PipelineId, String, usize, usize, String),
}

Messages sent from the constellation or layout to the script thread.

Variants
AttachLayout(NewLayoutInfo)

Gives a channel and ID to a layout thread, as well as the ID of that layout's parent

Resize(PipelineId, WindowSizeData, WindowSizeType)

Window resized. Sends a DOM event eventually, but first we combine events.

ResizeInactive(PipelineId, WindowSizeData)

Notifies script that window has been resized but to not take immediate action.

ExitPipeline(PipelineId)

Notifies the script that a pipeline should be closed.

SendEvent(PipelineId, CompositorEvent)

Sends a DOM event.

Viewport(PipelineId, Rect<f32>)

Notifies script of the viewport.

SetScrollState(PipelineId, Vec<(UntrustedNodeAddress, Point2D<f32>)>)

Notifies script of a new set of scroll offsets.

GetTitle(PipelineId)

Requests that the script thread immediately send the constellation the title of a pipeline.

Freeze(PipelineId)

Notifies script thread to suspend all its timers

Thaw(PipelineId)

Notifies script thread to resume all its timers

ChangeFrameVisibilityStatus(PipelineId, bool)

Notifies script thread whether frame is visible

NotifyVisibilityChange(PipelineId, PipelineId, bool)

Notifies script thread that frame visibility change is complete

Navigate(PipelineId, SubpageId, LoadData)

Notifies script thread that a url should be loaded in this iframe.

MozBrowserEvent(PipelineId, SubpageId, MozBrowserEvent)

Requests the script thread forward a mozbrowser event to an iframe it owns

UpdateSubpageId(PipelineId, SubpageId, SubpageId, PipelineId)

Updates the current subpage and pipeline IDs of a given iframe

FocusIFrame(PipelineId, SubpageId)

Set an iframe to be focused. Used when an element in an iframe gains focus.

WebDriverScriptCommand(PipelineId, WebDriverScriptCommand)

Passes a webdriver command to the script thread for execution

TickAllAnimations(PipelineId)

Notifies script thread that all animations are done

WebFontLoaded(PipelineId)

Notifies the script thread that a new Web font has been loaded, and thus the page should be reflowed.

DispatchFrameLoadEvent

Cause a load event to be dispatched at the appropriate frame element.

Fields

target: PipelineId

The pipeline that has been marked as loaded.

parent: PipelineId

The pipeline that contains a frame loading the target pipeline.

FramedContentChanged(PipelineId, SubpageId)

Notifies a parent frame that one of its child frames is now active.

ReportCSSError(PipelineId, String, usize, usize, String)

Report an error from a CSS parser for the given pipeline