Enum script_traits::ConstellationControlMsg

pub enum ConstellationControlMsg {
    AttachLayout(NewLayoutInfo),
    Resize(PipelineId, WindowSizeData, WindowSizeType),
    ResizeInactive(PipelineId, WindowSizeData),
    ExitPipeline(PipelineId),
    SendEvent(PipelineId, CompositorEvent),
    Viewport(PipelineId, Rect<f32>),
    GetTitle(PipelineId),
    Freeze(PipelineId),
    Thaw(PipelineId),
    Navigate(PipelineId, SubpageId, LoadData),
    MozBrowserEvent(PipelineId, SubpageId, MozBrowserEvent),
    UpdateSubpageId(PipelineId, SubpageId, SubpageId, PipelineId),
    FocusIFrame(PipelineId, SubpageId),
    WebDriverScriptCommand(PipelineId, WebDriverScriptCommand),
    TickAllAnimations(PipelineId),
    WebFontLoaded(PipelineId),
    DispatchFrameLoadEvent {
        target: PipelineId,
        parent: PipelineId,
    },
    FramedContentChanged(PipelineId, SubpageId),
    ReportCSSError(PipelineId, String, usize, usize, String),
}

Messages sent from the constellation or layout to the script thread.

Variants

AttachLayout

Gives a channel and ID to a layout thread, as well as the ID of that layout's parent

Resize

Window resized. Sends a DOM event eventually, but first we combine events.

ResizeInactive

Notifies script that window has been resized but to not take immediate action.

ExitPipeline

Notifies the script that a pipeline should be closed.

SendEvent

Sends a DOM event.

Viewport

Notifies script of the viewport.

GetTitle

Requests that the script thread immediately send the constellation the title of a pipeline.

Freeze

Notifies script thread to suspend all its timers

Thaw

Notifies script thread to resume all its timers

Navigate

Notifies script thread that a url should be loaded in this iframe.

MozBrowserEvent

Requests the script thread forward a mozbrowser event to an iframe it owns

UpdateSubpageId

Updates the current subpage and pipeline IDs of a given iframe

FocusIFrame

Set an iframe to be focused. Used when an element in an iframe gains focus.

WebDriverScriptCommand

Passes a webdriver command to the script thread for execution

TickAllAnimations

Notifies script thread that all animations are done

WebFontLoaded

Notifies the script thread that a new Web font has been loaded, and thus the page should be reflowed.

DispatchFrameLoadEvent

Cause a load event to be dispatched at the appropriate frame element.

Fields

target

The pipeline that has been marked as loaded.

parent

The pipeline that contains a frame loading the target pipeline.

FramedContentChanged

Notifies a parent frame that one of its child frames is now active.

ReportCSSError

Report an error from a CSS parser for the given pipeline