Enum script_traits::ConstellationControlMsg

pub enum ConstellationControlMsg {

    StopDelayingLoadEventsMode(PipelineId),
    NavigationResponse(PipelineId, FetchResponseMsg),
    AttachLayout(NewLayoutInfo),
    Resize(PipelineId, WindowSizeData, WindowSizeType),
    ResizeInactive(PipelineId, WindowSizeData),
    ExitFullScreen(PipelineId),
    UnloadDocument(PipelineId),
    ExitPipeline(PipelineId, DiscardBrowsingContext),
    ExitScriptThread,
    SendEvent(PipelineId, CompositorEvent),
    Viewport(PipelineId, Rect<f32, UnknownUnit>),
    SetScrollState(PipelineId, Vec<(UntrustedNodeAddress, Vector2D<f32, LayoutPixel>)>),
    GetTitle(PipelineId),
    SetDocumentActivity(PipelineId, DocumentActivity),
    ChangeFrameVisibilityStatus(PipelineId, bool),
    NotifyVisibilityChange(PipelineId, BrowsingContextId, bool),
    NavigateIframe(PipelineId, BrowsingContextId, LoadData, HistoryEntryReplacement),
    PostMessage {
        target: PipelineId,
        source: PipelineId,
        source_browsing_context: TopLevelBrowsingContextId,
        target_origin: Option<ImmutableOrigin>,
        source_origin: ImmutableOrigin,
        data: StructuredSerializedData,
    },
    UpdatePipelineId(PipelineId, BrowsingContextId, TopLevelBrowsingContextId, PipelineId, UpdatePipelineIdReason),
    UpdateHistoryState(PipelineId, Option<HistoryStateId>, ServoUrl),
    RemoveHistoryStates(PipelineId, Vec<HistoryStateId>),
    FocusIFrame(PipelineId, BrowsingContextId),
    WebDriverScriptCommand(PipelineId, WebDriverScriptCommand),
    TickAllAnimations(PipelineId, AnimationTickType),
    WebFontLoaded(PipelineId),
    DispatchIFrameLoadEvent {
        target: BrowsingContextId,
        parent: PipelineId,
        child: PipelineId,
    },
    DispatchStorageEvent(PipelineId, StorageType, ServoUrl, Option<String>, Option<String>, Option<String>),
    ReportCSSError(PipelineId, String, u32, u32, String),
    Reload(PipelineId),
    PaintMetric(PipelineId, ProgressiveWebMetricType, u64),
    MediaSessionAction(PipelineId, MediaSessionActionType),
    SetWebGPUPort(IpcReceiver<WebGPUMsg>),

}

Messages sent from the constellation or layout to the script thread.

Variants

StopDelayingLoadEventsMode

Takes the associated window proxy out of “delaying-load-events-mode”, used if a scheduled navigated was refused by the embedder. https://html.spec.whatwg.org/multipage/#delaying-load-events-mode

Tuple Fields of StopDelayingLoadEventsMode

0: PipelineId

NavigationResponse

Sends the final response to script thread for fetching after all redirections have been resolved

Tuple Fields of NavigationResponse

0: PipelineId1: FetchResponseMsg

AttachLayout

Gives a channel and ID to a layout thread, as well as the ID of that layout’s parent

Tuple Fields of AttachLayout

0: NewLayoutInfo

Resize

Window resized. Sends a DOM event eventually, but first we combine events.

Tuple Fields of Resize

0: PipelineId1: WindowSizeData2: WindowSizeType

ResizeInactive

Notifies script that window has been resized but to not take immediate action.

Tuple Fields of ResizeInactive

0: PipelineId1: WindowSizeData

ExitFullScreen

Window switched from fullscreen mode.

Tuple Fields of ExitFullScreen

0: PipelineId

UnloadDocument

Notifies the script that the document associated with this pipeline should ‘unload’.

Tuple Fields of UnloadDocument

0: PipelineId

ExitPipeline

Notifies the script that a pipeline should be closed.

Tuple Fields of ExitPipeline

0: PipelineId1: DiscardBrowsingContext

ExitScriptThread

Notifies the script that the whole thread should be closed.

SendEvent

Sends a DOM event.

Tuple Fields of SendEvent

0: PipelineId1: CompositorEvent

Viewport

Notifies script of the viewport.

Tuple Fields of Viewport

0: PipelineId1: Rect<f32, UnknownUnit>

SetScrollState

Notifies script of a new set of scroll offsets.

Tuple Fields of SetScrollState

0: PipelineId1: Vec<(UntrustedNodeAddress, Vector2D<f32, LayoutPixel>)>

GetTitle

Requests that the script thread immediately send the constellation the title of a pipeline.

Tuple Fields of GetTitle

0: PipelineId

SetDocumentActivity

Notifies script thread of a change to one of its document’s activity

Tuple Fields of SetDocumentActivity

0: PipelineId1: DocumentActivity

ChangeFrameVisibilityStatus

Notifies script thread whether frame is visible

Tuple Fields of ChangeFrameVisibilityStatus

0: PipelineId1: bool

NotifyVisibilityChange

Notifies script thread that frame visibility change is complete PipelineId is for the parent, BrowsingContextId is for the nested browsing context

Tuple Fields of NotifyVisibilityChange

0: PipelineId1: BrowsingContextId2: bool

NavigateIframe

Notifies script thread that a url should be loaded in this iframe. PipelineId is for the parent, BrowsingContextId is for the nested browsing context

Tuple Fields of NavigateIframe

0: PipelineId1: BrowsingContextId2: LoadData3: HistoryEntryReplacement

PostMessage

Post a message to a given window.

Fields of PostMessage

target: PipelineId

The target of the message.

source: PipelineId

The source of the message.

source_browsing_context: TopLevelBrowsingContextId

The top level browsing context associated with the source pipeline.

target_origin: Option<ImmutableOrigin>

The expected origin of the target.

source_origin: ImmutableOrigin

The source origin of the message. https://html.spec.whatwg.org/multipage/#dom-messageevent-origin

data: StructuredSerializedData

The data to be posted.

UpdatePipelineId

Updates the current pipeline ID of a given iframe. First PipelineId is for the parent, second is the new PipelineId for the frame.

Tuple Fields of UpdatePipelineId

0: PipelineId1: BrowsingContextId2: TopLevelBrowsingContextId3: PipelineId4: UpdatePipelineIdReason

UpdateHistoryState

Updates the history state and url of a given pipeline.

Tuple Fields of UpdateHistoryState

0: PipelineId1: Option<HistoryStateId>2: ServoUrl

RemoveHistoryStates

Removes inaccesible history states.

Tuple Fields of RemoveHistoryStates

0: PipelineId1: Vec<HistoryStateId>

FocusIFrame

Set an iframe to be focused. Used when an element in an iframe gains focus. PipelineId is for the parent, BrowsingContextId is for the nested browsing context

Tuple Fields of FocusIFrame

0: PipelineId1: BrowsingContextId

WebDriverScriptCommand

Passes a webdriver command to the script thread for execution

Tuple Fields of WebDriverScriptCommand

0: PipelineId1: WebDriverScriptCommand

TickAllAnimations

Notifies script thread that all animations are done

Tuple Fields of TickAllAnimations

0: PipelineId1: AnimationTickType

WebFontLoaded

Notifies the script thread that a new Web font has been loaded, and thus the page should be reflowed.

Tuple Fields of WebFontLoaded

0: PipelineId

DispatchIFrameLoadEvent

Cause a load event to be dispatched at the appropriate iframe element.

Fields of DispatchIFrameLoadEvent

target: BrowsingContextId

The frame that has been marked as loaded.

parent: PipelineId

The pipeline that contains a frame loading the target pipeline.

child: PipelineId

The pipeline that has completed loading.

DispatchStorageEvent

Cause a storage event to be dispatched at the appropriate window. The strings are key, old value and new value.

Tuple Fields of DispatchStorageEvent

0: PipelineId1: StorageType2: ServoUrl3: Option<String>4: Option<String>5: Option<String>

ReportCSSError

Report an error from a CSS parser for the given pipeline

Tuple Fields of ReportCSSError

0: PipelineId1: String2: u323: u324: String

Reload

Reload the given page.

Tuple Fields of Reload

0: PipelineId

PaintMetric

Notifies the script thread about a new recorded paint metric.

Tuple Fields of PaintMetric

0: PipelineId1: ProgressiveWebMetricType2: u64

MediaSessionAction

Notifies the media session about a user requested media session action.

Tuple Fields of MediaSessionAction

0: PipelineId1: MediaSessionActionType

SetWebGPUPort

Notifies script thread that WebGPU server has started

Tuple Fields of SetWebGPUPort

0: IpcReceiver<WebGPUMsg>