Enum script_traits::ConstellationControlMsg

pub enum ConstellationControlMsg {
    AttachLayout(NewLayoutInfo),
    Resize(PipelineId, WindowSizeData),
    ResizeInactive(PipelineId, WindowSizeData),
    ExitPipeline(PipelineId, PipelineExitType),
    SendEvent(PipelineId, CompositorEvent),
    ReflowComplete(PipelineId, u32),
    Viewport(PipelineId, Rect<f32>),
    GetTitle(PipelineId),
    Freeze(PipelineId),
    Thaw(PipelineId),
    Navigate(PipelineId, SubpageId, LoadData),
    MozBrowserEventMsg(PipelineId, SubpageId, MozBrowserEvent),
    UpdateSubpageId(PipelineId, SubpageId, SubpageId),
}

Messages sent from the constellation to the script task

Variants

AttachLayout

Gives a channel and ID to a layout task, as well as the ID of that layout's parent

Resize

Window resized. Sends a DOM event eventually, but first we combine events.

ResizeInactive

Notifies script that window has been resized but to not take immediate action.

ExitPipeline

Notifies the script that a pipeline should be closed.

SendEvent

Sends a DOM event.

ReflowComplete

Notifies script that reflow is finished.

Viewport

Notifies script of the viewport.

GetTitle

Requests that the script task immediately send the constellation the title of a pipeline.

Freeze

Notifies script task to suspend all its timers

Thaw

Notifies script task to resume all its timers

Navigate

Notifies script task that a url should be loaded in this iframe.

MozBrowserEventMsg

Requests the script task forward a mozbrowser event to an iframe it owns

UpdateSubpageId

Updates the current subpage id of a given iframe