Enum script_traits::MozBrowserEvent

pub enum MozBrowserEvent {
    AsyncScroll,
    Close,
    ContextMenu,
    Error(MozBrowserErrorType, Option<String>, Option<String>),
    IconChange(String, String, String),
    Connected,
    LoadEnd,
    LoadStart,
    LocationChange(String, bool, bool),
    OpenWindow,
    SecurityChange(HttpsState),
    ShowModalPrompt(String, String, String, String),
    TitleChange(String),
    UsernameAndPasswordRequired,
    OpenSearch,
}

The events fired in a Browser API context (<iframe mozbrowser>)

Variants

AsyncScroll

Sent when the scroll position within a browser <iframe> changes.

Close

Sent when window.close() is called within a browser <iframe>.

ContextMenu

Sent when a browser <iframe> tries to open a context menu. This allows handling <menuitem> element available within the browser <iframe>'s content.

Error

Sent when an error occurred while trying to load content within a browser <iframe>.

IconChange

Sent when the favicon of a browser <iframe> changes.

Connected

Sent when the browser <iframe> has reached the server.

LoadEnd

Sent when the browser <iframe> has finished loading all its assets.

LoadStart

Sent when the browser <iframe> starts to load a new page.

LocationChange

Sent when a browser <iframe>'s location changes.

OpenWindow

Sent when window.open() is called within a browser <iframe>.

SecurityChange

Sent when the SSL state changes within a browser <iframe>.

ShowModalPrompt

Sent when alert(), confirm(), or prompt() is called within a browser <iframe>.

TitleChange

Sent when the document.title changes within a browser <iframe>.

UsernameAndPasswordRequired

Sent when an HTTP authentification is requested.

OpenSearch

Sent when a link to a search engine is found.

Methods

impl MozBrowserEvent

fn name(&self) -> &'static str

Get the name of the event as a & str