Enum script_traits::ScriptMsg

pub enum ScriptMsg {
    ChangeRunningAnimationsState(PipelineId, AnimationState),
    CreateCanvasPaintThread(Size2D<i32>, IpcSender<IpcSender<CanvasMsg>>),
    CreateWebGLPaintThread(Size2D<i32>, GLContextAttributes, IpcSender<Result<(IpcSender<CanvasMsg>, GLLimits), String>>),
    DOMLoad(PipelineId),
    Focus(PipelineId),
    ForwardMouseButtonEvent(PipelineId, MouseEventType, MouseButton, Point2D<f32>),
    ForwardMouseMoveEvent(PipelineId, Point2D<f32>),
    GetClipboardContents(IpcSender<String>),
    HeadParsed,
    LoadComplete(PipelineId),
    LoadUrl(PipelineId, LoadData),
    MozBrowserEvent(PipelineId, SubpageId, MozBrowserEvent),
    Navigate(Option<(PipelineId, SubpageId)>, NavigationDirection),
    NewFavicon(Url),
    NodeStatus(Option<String>),
    RemoveIFrame(PipelineId, Option<IpcSender<()>>),
    ScriptLoadedURLInIFrame(IFrameLoadInfo),
    SetClipboardContents(String),
    ActivateDocument(PipelineId),
    SetDocumentState(PipelineId, DocumentState),
    SetFinalUrl(PipelineId, Url),
    Alert(PipelineId, String, IpcSender<bool>),
    ScrollFragmentPoint(PipelineId, LayerId, Point2D<f32>, bool),
    SetTitle(PipelineId, Option<String>),
    SendKeyEvent(Key, KeyState, KeyModifiers),
    GetClientWindow(IpcSender<(Size2D<u32>, Point2D<i32>)>),
    MoveTo(Point2D<i32>),
    ResizeTo(Size2D<u32>),
    TouchEventProcessed(EventResult),
    GetScrollOffset(PipelineId, LayerId, IpcSender<Point2D<f32>>),
    PipelineExited(PipelineId),
    Exit,
}

Messages from the script to the constellation.

Variants
ChangeRunningAnimationsState(PipelineId, AnimationState)

Indicates whether this pipeline is currently running animations.

CreateCanvasPaintThread(Size2D<i32>, IpcSender<IpcSender<CanvasMsg>>)

Requests that a new 2D canvas thread be created. (This is done in the constellation because 2D canvases may use the GPU and we don't want to give untrusted content access to the GPU.)

CreateWebGLPaintThread(Size2D<i32>, GLContextAttributes, IpcSender<Result<(IpcSender<CanvasMsg>, GLLimits), String>>)

Requests that a new WebGL thread be created. (This is done in the constellation because WebGL uses the GPU and we don't want to give untrusted content access to the GPU.)

DOMLoad(PipelineId)

Dispatched after the DOM load event has fired on a document Causes a load event to be dispatched to any enclosing frame context element for the given pipeline.

Focus(PipelineId)

Notifies the constellation that this frame has received focus.

ForwardMouseButtonEvent(PipelineId, MouseEventType, MouseButton, Point2D<f32>)

Re-send a mouse button event that was sent to the parent window.

ForwardMouseMoveEvent(PipelineId, Point2D<f32>)

Re-send a mouse move event that was sent to the parent window.

GetClipboardContents(IpcSender<String>)

Requests that the constellation retrieve the current contents of the clipboard

HeadParsed

LoadComplete(PipelineId)

All pending loads are complete.

LoadUrl(PipelineId, LoadData)

A new load has been requested.

MozBrowserEvent(PipelineId, SubpageId, MozBrowserEvent)

Dispatch a mozbrowser event to a given iframe. Only available in experimental mode.

Navigate(Option<(PipelineId, SubpageId)>, NavigationDirection)

HTMLIFrameElement Forward or Back navigation.

NewFavicon(Url)

Favicon detected

NodeStatus(Option<String>)

Status message to be displayed in the chrome, eg. a link URL on mouseover.

RemoveIFrame(PipelineId, Option<IpcSender<()>>)

Notification that this iframe should be removed.

ScriptLoadedURLInIFrame(IFrameLoadInfo)

A load has been requested in an IFrame.

SetClipboardContents(String)

Requests that the constellation set the contents of the clipboard

ActivateDocument(PipelineId)

Mark a new document as active

SetDocumentState(PipelineId, DocumentState)

Set the document state for a pipeline (used by screenshot / reftests)

SetFinalUrl(PipelineId, Url)

Update the pipeline Url, which can change after redirections.

Alert(PipelineId, String, IpcSender<bool>)

Check if an alert dialog box should be presented

ScrollFragmentPoint(PipelineId, LayerId, Point2D<f32>, bool)

Scroll a page in a window

SetTitle(PipelineId, Option<String>)

Set title of current page https://html.spec.whatwg.org/multipage/#document.title

SendKeyEvent(Key, KeyState, KeyModifiers)

Send a key event

GetClientWindow(IpcSender<(Size2D<u32>, Point2D<i32>)>)

Get Window Informations size and position

MoveTo(Point2D<i32>)

Move the window to a point

ResizeTo(Size2D<u32>)

Resize the window to size

TouchEventProcessed(EventResult)

Script has handled a touch event, and either prevented or allowed default actions.

GetScrollOffset(PipelineId, LayerId, IpcSender<Point2D<f32>>)

Get Scroll Offset

PipelineExited(PipelineId)

Notifies the constellation that this pipeline has exited.

Exit

Requests that the compositor shut down.