Enum script_traits::ScriptMsg

pub enum ScriptMsg {

    CompleteMessagePortTransfer(MessagePortRouterId, Vec<MessagePortId>),
    MessagePortTransferResult(Option<MessagePortRouterId>, Vec<MessagePortId>, HashMap<MessagePortId, VecDeque<PortMessageTask>>),
    NewMessagePort(MessagePortRouterId, MessagePortId),
    NewMessagePortRouter(MessagePortRouterId, IpcSender<MessagePortMsg>),
    RemoveMessagePortRouter(MessagePortRouterId),
    RerouteMessagePort(MessagePortId, PortMessageTask),
    MessagePortShipped(MessagePortId),
    RemoveMessagePort(MessagePortId),
    EntanglePorts(MessagePortId, MessagePortId),
    NewBroadcastChannelRouter(BroadcastChannelRouterId, IpcSender<BroadcastMsg>, ImmutableOrigin),
    RemoveBroadcastChannelRouter(BroadcastChannelRouterId, ImmutableOrigin),
    NewBroadcastChannelNameInRouter(BroadcastChannelRouterId, String, ImmutableOrigin),
    RemoveBroadcastChannelNameInRouter(BroadcastChannelRouterId, String, ImmutableOrigin),
    ScheduleBroadcast(BroadcastChannelRouterId, BroadcastMsg),
    ForwardToEmbedder(EmbedderMsg),
    InitiateNavigateRequest(RequestBuilder, IpcReceiver<()>),
    BroadcastStorageEvent(StorageType, ServoUrl, Option<String>, Option<String>, Option<String>),
    ChangeRunningAnimationsState(AnimationState),
    CreateCanvasPaintThread(UntypedSize2D<u64>, IpcSender<(IpcSender<CanvasMsg>, CanvasId)>),
    Focus,
    GetTopForBrowsingContext(BrowsingContextId, IpcSender<Option<TopLevelBrowsingContextId>>),
    GetBrowsingContextInfo(PipelineId, IpcSender<Option<(BrowsingContextId, Option<PipelineId>)>>),
    GetChildBrowsingContextId(BrowsingContextId, usize, IpcSender<Option<BrowsingContextId>>),
    LoadComplete,
    LoadUrl(LoadData, HistoryEntryReplacement),
    AbortLoadUrl,
    PostMessage {
        target: BrowsingContextId,
        source: PipelineId,
        target_origin: Option<ImmutableOrigin>,
        source_origin: ImmutableOrigin,
        data: StructuredSerializedData,
    },
    NavigatedToFragment(ServoUrl, HistoryEntryReplacement),
    TraverseHistory(TraversalDirection),
    PushHistoryState(HistoryStateId, ServoUrl),
    ReplaceHistoryState(HistoryStateId, ServoUrl),
    JointSessionHistoryLength(IpcSender<u32>),
    RemoveIFrame(BrowsingContextId, IpcSender<Vec<PipelineId>>),
    VisibilityChangeComplete(bool),
    ScriptLoadedURLInIFrame(IFrameLoadInfoWithData),
    ScriptNewIFrame(IFrameLoadInfoWithData, IpcSender<LayoutControlMsg>),
    ScriptNewAuxiliary(AuxiliaryBrowsingContextLoadInfo, IpcSender<LayoutControlMsg>),
    ActivateDocument,
    SetDocumentState(DocumentState),
    SetFinalUrl(ServoUrl),
    TouchEventProcessed(EventResult),
    LogEntry(Option<String>, LogEntry),
    DiscardDocument,
    DiscardTopLevelBrowsingContext,
    PipelineExited,
    ForwardDOMMessage(DOMMessage, ServoUrl),
    ScheduleJob(Job),
    GetClientWindow(IpcSender<(DeviceIntSize, DeviceIntPoint)>),
    GetScreenSize(IpcSender<DeviceIntSize>),
    GetScreenAvailSize(IpcSender<DeviceIntSize>),
    MediaSessionEvent(PipelineId, MediaSessionEvent),
    RequestAdapter(IpcSender<WebGPUResponseResult>, RequestAdapterOptions, SmallVec<[AdapterId; 4]>),
    GetWebGPUChan(IpcSender<WebGPU>),
    TitleChanged(PipelineId, String),

}

Messages from the script to the constellation.

Variants

CompleteMessagePortTransfer

Request to complete the transfer of a set of ports to a router.

Tuple Fields of CompleteMessagePortTransfer

0: MessagePortRouterId1: Vec<MessagePortId>

MessagePortTransferResult

The results of attempting to complete the transfer of a batch of ports.

Tuple Fields of MessagePortTransferResult

0: Option<MessagePortRouterId>1: Vec<MessagePortId>2: HashMap<MessagePortId, VecDeque<PortMessageTask>>

NewMessagePort

A new message-port was created or transferred, with corresponding control-sender.

Tuple Fields of NewMessagePort

0: MessagePortRouterId1: MessagePortId

NewMessagePortRouter

A global has started managing message-ports

Tuple Fields of NewMessagePortRouter

0: MessagePortRouterId1: IpcSender<MessagePortMsg>

RemoveMessagePortRouter

A global has stopped managing message-ports

Tuple Fields of RemoveMessagePortRouter

0: MessagePortRouterId

RerouteMessagePort

A task requires re-routing to an already shipped message-port.

Tuple Fields of RerouteMessagePort

0: MessagePortId1: PortMessageTask

MessagePortShipped

A message-port was shipped, let the entangled port know.

Tuple Fields of MessagePortShipped

0: MessagePortId

RemoveMessagePort

A message-port has been discarded by script.

Tuple Fields of RemoveMessagePort

0: MessagePortId

EntanglePorts

Entangle two message-ports.

Tuple Fields of EntanglePorts

0: MessagePortId1: MessagePortId

NewBroadcastChannelRouter

A global has started managing broadcast-channels.

Tuple Fields of NewBroadcastChannelRouter

0: BroadcastChannelRouterId1: IpcSender<BroadcastMsg>2: ImmutableOrigin

RemoveBroadcastChannelRouter

A global has stopped managing broadcast-channels.

Tuple Fields of RemoveBroadcastChannelRouter

0: BroadcastChannelRouterId1: ImmutableOrigin

NewBroadcastChannelNameInRouter

A global started managing broadcast channels for a given channel-name.

Tuple Fields of NewBroadcastChannelNameInRouter

0: BroadcastChannelRouterId1: String2: ImmutableOrigin

RemoveBroadcastChannelNameInRouter

A global stopped managing broadcast channels for a given channel-name.

Tuple Fields of RemoveBroadcastChannelNameInRouter

0: BroadcastChannelRouterId1: String2: ImmutableOrigin

ScheduleBroadcast

Broadcast a message to all same-origin broadcast channels, excluding the source of the broadcast.

Tuple Fields of ScheduleBroadcast

0: BroadcastChannelRouterId1: BroadcastMsg

ForwardToEmbedder

Forward a message to the embedder.

Tuple Fields of ForwardToEmbedder

0: EmbedderMsg

InitiateNavigateRequest

Requests are sent to constellation and fetches are checked manually for cross-origin loads

Tuple Fields of InitiateNavigateRequest

0: RequestBuilder1: IpcReceiver<()>

BroadcastStorageEvent

Broadcast a storage event to every same-origin pipeline. The strings are key, old value and new value.

Tuple Fields of BroadcastStorageEvent

0: StorageType1: ServoUrl2: Option<String>3: Option<String>4: Option<String>

ChangeRunningAnimationsState

Indicates whether this pipeline is currently running animations.

Tuple Fields of ChangeRunningAnimationsState

0: AnimationState

CreateCanvasPaintThread

Requests that a new 2D canvas thread be created. (This is done in the constellation because 2D canvases may use the GPU and we don’t want to give untrusted content access to the GPU.)

Tuple Fields of CreateCanvasPaintThread

0: UntypedSize2D<u64>1: IpcSender<(IpcSender<CanvasMsg>, CanvasId)>

Focus

Notifies the constellation that this frame has received focus.

GetTopForBrowsingContext

Get the top-level browsing context info for a given browsing context.

Tuple Fields of GetTopForBrowsingContext

0: BrowsingContextId1: IpcSender<Option<TopLevelBrowsingContextId>>

GetBrowsingContextInfo

Get the browsing context id of the browsing context in which pipeline is embedded and the parent pipeline id of that browsing context.

Tuple Fields of GetBrowsingContextInfo

0: PipelineId1: IpcSender<Option<(BrowsingContextId, Option<PipelineId>)>>

GetChildBrowsingContextId

Get the nth child browsing context ID for a given browsing context, sorted in tree order.

Tuple Fields of GetChildBrowsingContextId

0: BrowsingContextId1: usize2: IpcSender<Option<BrowsingContextId>>

LoadComplete

All pending loads are complete, and the load event for this pipeline has been dispatched.

LoadUrl

A new load has been requested, with an option to replace the current entry once loaded instead of adding a new entry.

Tuple Fields of LoadUrl

0: LoadData1: HistoryEntryReplacement

AbortLoadUrl

Abort loading after sending a LoadUrl message.

PostMessage

Post a message to the currently active window of a given browsing context.

Fields of PostMessage

target: BrowsingContextId

The target of the posted message.

source: PipelineId

The source of the posted message.

target_origin: Option<ImmutableOrigin>

The expected origin of the target.

source_origin: ImmutableOrigin

The source origin of the message. https://html.spec.whatwg.org/multipage/#dom-messageevent-origin

data: StructuredSerializedData

The data to be posted.

NavigatedToFragment

Inform the constellation that a fragment was navigated to and whether or not it was a replacement navigation.

Tuple Fields of NavigatedToFragment

0: ServoUrl1: HistoryEntryReplacement

TraverseHistory

HTMLIFrameElement Forward or Back traversal.

Tuple Fields of TraverseHistory

0: TraversalDirection

PushHistoryState

Inform the constellation of a pushed history state.

Tuple Fields of PushHistoryState

0: HistoryStateId1: ServoUrl

ReplaceHistoryState

Inform the constellation of a replaced history state.

Tuple Fields of ReplaceHistoryState

0: HistoryStateId1: ServoUrl

JointSessionHistoryLength

Gets the length of the joint session history from the constellation.

Tuple Fields of JointSessionHistoryLength

0: IpcSender<u32>

RemoveIFrame

Notification that this iframe should be removed. Returns a list of pipelines which were closed.

Tuple Fields of RemoveIFrame

0: BrowsingContextId1: IpcSender<Vec<PipelineId>>

VisibilityChangeComplete

Notifies constellation that an iframe’s visibility has been changed.

Tuple Fields of VisibilityChangeComplete

0: bool

ScriptLoadedURLInIFrame

A load has been requested in an IFrame.

Tuple Fields of ScriptLoadedURLInIFrame

0: IFrameLoadInfoWithData

ScriptNewIFrame

A load of the initial about:blank has been completed in an IFrame.

Tuple Fields of ScriptNewIFrame

0: IFrameLoadInfoWithData1: IpcSender<LayoutControlMsg>

ScriptNewAuxiliary

Script has opened a new auxiliary browsing context.

Tuple Fields of ScriptNewAuxiliary

0: AuxiliaryBrowsingContextLoadInfo1: IpcSender<LayoutControlMsg>

ActivateDocument

Mark a new document as active

SetDocumentState

Set the document state for a pipeline (used by screenshot / reftests)

Tuple Fields of SetDocumentState

0: DocumentState

SetFinalUrl

Update the pipeline Url, which can change after redirections.

Tuple Fields of SetFinalUrl

0: ServoUrl

TouchEventProcessed

Script has handled a touch event, and either prevented or allowed default actions.

Tuple Fields of TouchEventProcessed

0: EventResult

LogEntry

A log entry, with the top-level browsing context id and thread name

Tuple Fields of LogEntry

0: Option<String>1: LogEntry

DiscardDocument

Discard the document.

DiscardTopLevelBrowsingContext

Discard the browsing context.

PipelineExited

Notifies the constellation that this pipeline has exited.

ForwardDOMMessage

Send messages from postMessage calls from serviceworker to constellation for storing in service worker manager

Tuple Fields of ForwardDOMMessage

0: DOMMessage1: ServoUrl

ScheduleJob

https://w3c.github.io/ServiceWorker/#schedule-job-algorithm.

Tuple Fields of ScheduleJob

0: Job

GetClientWindow

Get Window Informations size and position

Tuple Fields of GetClientWindow

0: IpcSender<(DeviceIntSize, DeviceIntPoint)>

GetScreenSize

Get the screen size (pixel)

Tuple Fields of GetScreenSize

0: IpcSender<DeviceIntSize>

GetScreenAvailSize

Get the available screen size (pixel)

Tuple Fields of GetScreenAvailSize

0: IpcSender<DeviceIntSize>

MediaSessionEvent

Notifies the constellation about media session events (i.e. when there is metadata for the active media session, playback state changes…).

Tuple Fields of MediaSessionEvent

0: PipelineId1: MediaSessionEvent

RequestAdapter

Create a WebGPU Adapter instance

Tuple Fields of RequestAdapter

0: IpcSender<WebGPUResponseResult>1: RequestAdapterOptions2: SmallVec<[AdapterId; 4]>

GetWebGPUChan

Get WebGPU channel

Tuple Fields of GetWebGPUChan

0: IpcSender<WebGPU>

TitleChanged

Notify the constellation of a pipeline’s document’s title.

Tuple Fields of TitleChanged

0: PipelineId1: String