Module script_traits::script_msg

Structs

DOMMessage
Message that gets passed to service worker scope on postMessage

IFrameSize
A particular iframe's size, associated with a browsing context.

IFrameSizeMsg
An iframe sizing operation.

SWManagerSenders
Channels to allow service worker manager to communicate with constellation and resource thread

ScopeThings
Entities required to spawn service workers

Enums

EventResult
Whether a DOM event was prevented by web content

HistoryEntryReplacement
https://html.spec.whatwg.org/multipage/#replacement-enabled

LayoutMsg
Messages from the layout to the constellation.

LogEntry
A log entry reported to the constellation We don't report all log entries, just serious ones. We need a separate type for this because LogLevel isn't serializable.

SWManagerMsg
Messages outgoing from the Service Worker Manager thread to constellation

ScriptMsg
Messages from the script to the constellation.

ServiceWorkerMsg
Messages sent to Service Worker Manager thread

Constants

_IMPL_DESERIALIZE_FOR_DOMMessage

_IMPL_DESERIALIZE_FOR_EventResult

_IMPL_DESERIALIZE_FOR_HistoryEntryReplacement

_IMPL_DESERIALIZE_FOR_IFrameSize

_IMPL_DESERIALIZE_FOR_IFrameSizeMsg

_IMPL_DESERIALIZE_FOR_LayoutMsg

_IMPL_DESERIALIZE_FOR_LogEntry

_IMPL_DESERIALIZE_FOR_SWManagerMsg

_IMPL_DESERIALIZE_FOR_ScopeThings

_IMPL_DESERIALIZE_FOR_ScriptMsg

_IMPL_DESERIALIZE_FOR_ServiceWorkerMsg

_IMPL_SERIALIZE_FOR_DOMMessage

_IMPL_SERIALIZE_FOR_EventResult

_IMPL_SERIALIZE_FOR_HistoryEntryReplacement

_IMPL_SERIALIZE_FOR_IFrameSize

_IMPL_SERIALIZE_FOR_IFrameSizeMsg

_IMPL_SERIALIZE_FOR_LayoutMsg

_IMPL_SERIALIZE_FOR_LogEntry

_IMPL_SERIALIZE_FOR_SWManagerMsg

_IMPL_SERIALIZE_FOR_ScopeThings

_IMPL_SERIALIZE_FOR_ScriptMsg

_IMPL_SERIALIZE_FOR_ServiceWorkerMsg