Enum script_traits::script_msg::ScriptMsg

pub enum ScriptMsg {
    ChangeRunningAnimationsState(PipelineId, AnimationState),
    CreateCanvasPaintThread(Size2D<i32>, IpcSender<IpcSender<CanvasMsg>>),
    CreateWebGLPaintThread(Size2D<i32>, GLContextAttributes, IpcSender<Result<(IpcSender<CanvasMsg>, GLLimits), String>>),
    DOMLoad(PipelineId),
    Focus(PipelineId),
    ForwardMouseButtonEvent(PipelineId, MouseEventType, MouseButton, Point2D<f32>),
    ForwardMouseMoveEvent(PipelineId, Point2D<f32>),
    GetClipboardContents(IpcSender<String>),
    HeadParsed,
    LoadComplete(PipelineId),
    LoadUrl(PipelineId, LoadData),
    MozBrowserEvent(PipelineId, SubpageId, MozBrowserEvent),
    Navigate(Option<(PipelineId, SubpageId)>, NavigationDirection),
    NewFavicon(Url),
    NodeStatus(Option<String>),
    RemoveIFrame(PipelineId, Option<IpcSender<()>>),
    ScriptLoadedURLInIFrame(IFrameLoadInfo),
    SetClipboardContents(String),
    ActivateDocument(PipelineId),
    SetDocumentState(PipelineId, DocumentState),
    SetFinalUrl(PipelineId, Url),
    Alert(PipelineId, String, IpcSender<bool>),
    ScrollFragmentPoint(PipelineId, LayerId, Point2D<f32>, bool),
    SetTitle(PipelineId, Option<String>),
    SendKeyEvent(Key, KeyState, KeyModifiers),
    GetClientWindow(IpcSender<(Size2D<u32>, Point2D<i32>)>),
    MoveTo(Point2D<i32>),
    ResizeTo(Size2D<u32>),
    TouchEventProcessed(EventResult),
    GetScrollOffset(PipelineId, LayerId, IpcSender<Point2D<f32>>),
    PipelineExited(PipelineId),
    Exit,
}

Messages from the script to the constellation.

Variants

ChangeRunningAnimationsState

Indicates whether this pipeline is currently running animations.

CreateCanvasPaintThread

Requests that a new 2D canvas thread be created. (This is done in the constellation because 2D canvases may use the GPU and we don't want to give untrusted content access to the GPU.)

CreateWebGLPaintThread

Requests that a new WebGL thread be created. (This is done in the constellation because WebGL uses the GPU and we don't want to give untrusted content access to the GPU.)

DOMLoad

Dispatched after the DOM load event has fired on a document Causes a load event to be dispatched to any enclosing frame context element for the given pipeline.

Focus

Notifies the constellation that this frame has received focus.

ForwardMouseButtonEvent

Re-send a mouse button event that was sent to the parent window.

ForwardMouseMoveEvent

Re-send a mouse move event that was sent to the parent window.

GetClipboardContents

Requests that the constellation retrieve the current contents of the clipboard

HeadParsed

LoadComplete

All pending loads are complete.

LoadUrl

A new load has been requested.

MozBrowserEvent

Dispatch a mozbrowser event to a given iframe. Only available in experimental mode.

Navigate

HTMLIFrameElement Forward or Back navigation.

NewFavicon

Favicon detected

NodeStatus

Status message to be displayed in the chrome, eg. a link URL on mouseover.

RemoveIFrame

Notification that this iframe should be removed.

ScriptLoadedURLInIFrame

A load has been requested in an IFrame.

SetClipboardContents

Requests that the constellation set the contents of the clipboard

ActivateDocument

Mark a new document as active

SetDocumentState

Set the document state for a pipeline (used by screenshot / reftests)

SetFinalUrl

Update the pipeline Url, which can change after redirections.

Alert

Check if an alert dialog box should be presented

ScrollFragmentPoint

Scroll a page in a window

SetTitle

Set title of current page https://html.spec.whatwg.org/multipage/#document.title

SendKeyEvent

Send a key event

GetClientWindow

Get Window Informations size and position

MoveTo

Move the window to a point

ResizeTo

Resize the window to size

TouchEventProcessed

Script has handled a touch event, and either prevented or allowed default actions.

GetScrollOffset

Get Scroll Offset

PipelineExited

Notifies the constellation that this pipeline has exited.

Exit

Requests that the compositor shut down.