Struct script_traits::InitialScriptState

pub struct InitialScriptState {
    pub id: PipelineId,
    pub parent_info: Option<(PipelineId, SubpageId, FrameType)>,
    pub control_chan: IpcSender<ConstellationControlMsg>,
    pub control_port: IpcReceiver<ConstellationControlMsg>,
    pub constellation_chan: IpcSender<ScriptMsg>,
    pub panic_chan: IpcSender<PanicMsg>,
    pub scheduler_chan: IpcSender<TimerEventRequest>,
    pub resource_threads: ResourceThreads,
    pub bluetooth_thread: IpcSender<BluetoothMethodMsg>,
    pub image_cache_thread: ImageCacheThread,
    pub time_profiler_chan: ProfilerChan,
    pub mem_profiler_chan: ProfilerChan,
    pub devtools_chan: Option<IpcSender<ScriptToDevtoolsControlMsg>>,
    pub window_size: Option<WindowSizeData>,
    pub pipeline_namespace_id: PipelineNamespaceId,
    pub content_process_shutdown_chan: IpcSender<()>,
}

Data needed to construct a script thread.

NB: DO NOT add any Senders or Receivers here! pcwalton will have to rewrite your code if you do! Use IPC senders and receivers instead.

Fields
id: PipelineId

The ID of the pipeline with which this script thread is associated.

parent_info: Option<(PipelineId, SubpageId, FrameType)>

The subpage ID of this pipeline to create in its pipeline parent. If None, this is the root.

control_chan: IpcSender<ConstellationControlMsg>

A channel with which messages can be sent to us (the script thread).

control_port: IpcReceiver<ConstellationControlMsg>

A port on which messages sent by the constellation to script can be received.

constellation_chan: IpcSender<ScriptMsg>

A channel on which messages can be sent to the constellation from script.

panic_chan: IpcSender<PanicMsg>

A channel for sending panics to the constellation.

scheduler_chan: IpcSender<TimerEventRequest>

A channel to schedule timer events.

resource_threads: ResourceThreads

A channel to the resource manager thread.

bluetooth_thread: IpcSender<BluetoothMethodMsg>

A channel to the bluetooth thread.

image_cache_thread: ImageCacheThread

A channel to the image cache thread.

time_profiler_chan: ProfilerChan

A channel to the time profiler thread.

mem_profiler_chan: ProfilerChan

A channel to the memory profiler thread.

devtools_chan: Option<IpcSender<ScriptToDevtoolsControlMsg>>

A channel to the developer tools, if applicable.

window_size: Option<WindowSizeData>

Information about the initial window size.

pipeline_namespace_id: PipelineNamespaceId

The ID of the pipeline namespace for this script thread.

content_process_shutdown_chan: IpcSender<()>

A ping will be sent on this channel once the script thread shuts down.