Struct script_traits::InitialScriptState

pub struct InitialScriptState {
    pub id: PipelineId,
    pub parent_info: Option<(PipelineId, SubpageId)>,
    pub compositor: IpcSender<ScriptToCompositorMsg>,
    pub control_chan: IpcSender<ConstellationControlMsg>,
    pub control_port: IpcReceiver<ConstellationControlMsg>,
    pub constellation_chan: ConstellationChan<ScriptMsg>,
    pub layout_to_constellation_chan: ConstellationChan<LayoutMsg>,
    pub panic_chan: ConstellationChan<PanicMsg>,
    pub scheduler_chan: IpcSender<TimerEventRequest>,
    pub resource_thread: ResourceThread,
    pub bluetooth_thread: IpcSender<BluetoothMethodMsg>,
    pub storage_thread: StorageThread,
    pub image_cache_thread: ImageCacheThread,
    pub time_profiler_chan: ProfilerChan,
    pub mem_profiler_chan: ProfilerChan,
    pub devtools_chan: Option<IpcSender<ScriptToDevtoolsControlMsg>>,
    pub window_size: Option<WindowSizeData>,
    pub pipeline_namespace_id: PipelineNamespaceId,
    pub content_process_shutdown_chan: IpcSender<()>,
}

Data needed to construct a script thread.

NB: DO NOT add any Senders or Receivers here! pcwalton will have to rewrite your code if you do! Use IPC senders and receivers instead.

Fields

id

The ID of the pipeline with which this script thread is associated.

parent_info

The subpage ID of this pipeline to create in its pipeline parent. If None, this is the root.

compositor

The compositor.

control_chan

A channel with which messages can be sent to us (the script thread).

control_port

A port on which messages sent by the constellation to script can be received.

constellation_chan

A channel on which messages can be sent to the constellation from script.

layout_to_constellation_chan

A channel for the layout thread to send messages to the constellation.

panic_chan

A channel for sending panics to the constellation.

scheduler_chan

A channel to schedule timer events.

resource_thread

A channel to the resource manager thread.

bluetooth_thread

A channel to the bluetooth thread.

storage_thread

A channel to the storage thread.

image_cache_thread

A channel to the image cache thread.

time_profiler_chan

A channel to the time profiler thread.

mem_profiler_chan

A channel to the memory profiler thread.

devtools_chan

A channel to the developer tools, if applicable.

window_size

Information about the initial window size.

pipeline_namespace_id

The ID of the pipeline namespace for this script thread.

content_process_shutdown_chan

A ping will be sent on this channel once the script thread shuts down.