Struct script_traits::NewLayoutInfo

pub struct NewLayoutInfo {
    pub containing_pipeline_id: PipelineId,
    pub new_pipeline_id: PipelineId,
    pub subpage_id: SubpageId,
    pub load_data: LoadData,
    pub paint_chan: OptionalOpaqueIpcSender,
    pub pipeline_port: IpcReceiver<LayoutControlMsg>,
    pub panic_chan: IpcSender<PanicMsg>,
    pub layout_shutdown_chan: IpcSender<()>,
    pub content_process_shutdown_chan: IpcSender<()>,
}

The initial data associated with a newly-created framed pipeline.

Fields

containing_pipeline_id

Id of the parent of this new pipeline.

new_pipeline_id

Id of the newly-created pipeline.

subpage_id

Id of the new frame associated with this pipeline.

load_data

Network request data which will be initiated by the script thread.

paint_chan

The paint channel, cast to OptionalOpaqueIpcSender. This is really an Sender<LayoutToPaintMsg>.

pipeline_port

A port on which layout can receive messages from the pipeline.

panic_chan

A channel for sending panics on

layout_shutdown_chan

A shutdown channel so that layout can notify others when it's done.

content_process_shutdown_chan

A shutdown channel so that layout can tell the content process to shut down when it's done.